Added GraphViewModel.
