+++ native/KotlinIOS/KotlinIOS/Views/GraphView.swift
@@ -9,37 +9,6 @@
 import SwiftUI
 import SharedCode
 
-class GraphProgressModel : ObservableObject {
-    var inProgress = true
-}
-
-public class GraphViewModel : ObservableObject {
-    
-    @Published var graphModel: GraphModel?
-    @Published var dataRequestInProgress = ProgressModel()
-    @Published var inProgress : Bool = true
-    
-    private let repository: IdeaRepository?
-    private let ideaModel: IdeaModel?
-    
-    init(repository: IdeaRepository, ideaModel: IdeaModel) {
-        self.repository = repository
-        self.ideaModel = ideaModel
-        fetchKotlin()
-    }
-    
-    func fetchKotlin() {
-        dataRequestInProgress.inProgress = true
-        
-        if let ideaId = ideaModel?.id {
-            repository?.fetchGraph(ideaId: ideaId, success: { data in
-                self.graphModel = data
-                self.dataRequestInProgress.inProgress = false
-            })
-        }
-    }
-}
-
 struct GraphView: View {
     var ideaModel: IdeaModel
     var ideaModelLogicDecorator: IdeaModelLogicDecorator
